requirements.txt
annotated-types==0.7.0
anyio==4.8.0
av==14.1.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
customtkinter==5.2.2
darkdetect==0.8.0
decord==0.6.0
fastapi==0.115.8
filelock==3.17.0
fsspec==2025.2.0
h11==0.14.0
huggingface-hub==0.28.1
idna==3.10
Jinja2==3.1.5
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.2
packaging==24.2
pi==0.1.2
pillow==11.1.0
pydantic==2.10.6
pydantic_core==2.27.2
python-dotenv==1.0.1
python-multipart==0.0.20
Python_IO==0.3
PyYAML==6.0.2
qwen-vl-utils==0.0.8
regex==2024.11.6
requests==2.32.3
safetensors==0.5.2
sentencepiece==0.2.0
setuptools==75.8.0
sniffio==1.3.1
starlette==0.45.3
sympy==1.13.1
tokenizers==0.21.0
torch==2.6.0
tqdm==4.67.1
transformers==4.48.2
typing_extensions==4.12.2
urllib3==2.3.0
uvicorn==0.34.0
runpod
bitsandbytes
accelerate
huggingface_hub
python-dotenv